+++ backend/src/main/java/com/christmas/recommend/service/LocationService.java
@@ -4,6 +4,7 @@ import static com.christmas.recommend.domain.RecommendKeyword.getKeywords;
 
 import com.christmas.infrastructure.image.ImageApiManager;
 import com.christmas.infrastructure.route.domain.PointType;
+import com.christmas.infrastructure.route.dto.RouteConditionDto;
 import com.christmas.infrastructure.route.dto.RouteInfo;
 import com.christmas.infrastructure.route.dto.XY;
 import com.christmas.infrastructure.route.service.RouteApiManager;
@@ -60,7 +61,25 @@ public class LocationService {
         return searchApiManager.findLocationsByCategory(condition);
     }
 
-    public List<RouteInfo> getRoutes(List<XY> courseRoute, JsonNode routesDistance) {
+    public List<RouteInfo> findPedestrianRoute(List<XY> courseRoute) {
+        RouteConditionDto condition = makeRouteCondition(courseRoute);
+        JsonNode routesRaw = routeApiManager.getPedestrianRoute(condition);
+        return getRoutes(courseRoute, routesRaw);
+    }
+
+    private RouteConditionDto makeRouteCondition(List<XY> locationsNotNull) {
+        if (locationsNotNull.size() == 2) {
+            return new RouteConditionDto(locationsNotNull.get(0), "시작 장소", locationsNotNull.get(1), "종료 장소", List.of());
+        }
+        List<XY> passPlace = new ArrayList<>();
+        for (int i = 1; i < locationsNotNull.size() - 1; i++) {
+            passPlace.add(locationsNotNull.get(i));
+        }
+        return new RouteConditionDto(locationsNotNull.get(0), "시작 장소",
+                locationsNotNull.get(locationsNotNull.size() - 1), "종료 장소", passPlace);
+    }
+
+    private List<RouteInfo> getRoutes(List<XY> courseRoute, JsonNode routesDistance) {
         RouteApiParser parser = RouteApiParser.from(routesDistance);
         if (courseRoute.size() == 2) {
             return List.of(parser.getDistanceInfo(PointType.SP, PointType.EP));
@@ -80,6 +99,10 @@ public class LocationService {
             }
         }
         return routeInfos;
+    }
+
+    public String findPlaceImage(String placeName, XY xy) {
+        return imageApiManager.findPlaceImage(placeName, xy.x(), xy.y());
     }
 
     private SearchConditionDto setConditionByKeyword(CourseGetRequest request, LocationCategory category) {
