+++ backend/src/main/java/com/christmas/recommend/service/RecommendService.java
@@ -1,12 +1,8 @@
 package com.christmas.recommend.service;
 
-import com.christmas.infrastructure.image.ImageApiManager;
-import com.christmas.infrastructure.route.dto.RouteConditionDto;
 import com.christmas.infrastructure.route.dto.RouteInfo;
 import com.christmas.infrastructure.route.dto.XY;
-import com.christmas.infrastructure.route.service.RouteApiManager;
 import com.christmas.infrastructure.search.domain.LocationCategory;
-import com.christmas.infrastructure.search.service.SearchApiManager;
 import com.christmas.infrastructure.search.service.SearchApiParser;
 import com.christmas.recommend.domain.Course;
 import com.christmas.recommend.domain.Location;
@@ -17,7 +13,6 @@ import com.christmas.recommend.dto.CourseGetResponse;
 import com.christmas.util.RandomIntPicker;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -30,11 +25,8 @@ public class RecommendService {
 
     public static final int RECOMMEND_ATTRACTION_COUNT = 3;
 
-    private final SearchApiManager searchApiManager;
-    private final RouteApiManager routeApiManager;
     private final SearchApiParser searchApiParser;
     private final RandomIntPicker randomPicker;
-    private final ImageApiManager imageApiManager;
     private final LocationService locationService;
 
     public CourseGetResponse generateCourse(CourseGetRequest request) {
@@ -50,10 +42,7 @@ public class RecommendService {
         if (!course.isExist()) {
             return new CourseGetResponse(null, null, null, null);
         }
-        RouteConditionDto condition = makeRouteCondition(courseRoute);
-        JsonNode routesRaw = routeApiManager.getPedestrianRoute(condition);
-        List<RouteInfo> routes = locationService.getRoutes(courseRoute, routesRaw);
-
+        List<RouteInfo> routes = locationService.findPedestrianRoute(courseRoute);
         course.addRouteInfo(routes);
 
         Location lunch = course.getLunch();
@@ -78,7 +67,7 @@ public class RecommendService {
         if (location.isExist()) {
             String placeName = searchApiParser.extractNameFromLocation(location.getRaw());
             XY xy = searchApiParser.extractXYFromLocation(location.getRaw());
-            return imageApiManager.findPlaceImage(placeName, xy.x(), xy.y());
+            return locationService.findPlaceImage(placeName, xy);
         }
         return null;
     }
@@ -87,25 +76,13 @@ public class RecommendService {
         location.putField("image_url", imageUrl);
     }
 
-    private RouteConditionDto makeRouteCondition(List<XY> locationsNotNull) {
-        if (locationsNotNull.size() == 2) {
-            return new RouteConditionDto(locationsNotNull.get(0), "시작 장소", locationsNotNull.get(1), "종료 장소", List.of());
-        }
-        List<XY> passPlace = new ArrayList<>();
-        for (int i = 1; i < locationsNotNull.size() - 1; i++) {
-            passPlace.add(locationsNotNull.get(i));
-        }
-        return new RouteConditionDto(locationsNotNull.get(0), "시작 장소",
-                locationsNotNull.get(locationsNotNull.size() - 1), "종료 장소", passPlace);
-    }
-
     public AttractionGetResponse generateAttractions(AttractionGetRequest request) {
         List<JsonNode> attractions = locationService.findLocationsByCategory(request, LocationCategory.CULTURE);
         List<JsonNode> randomAttractions = getRandomLocations(attractions, RECOMMEND_ATTRACTION_COUNT);
         for (JsonNode attraction : randomAttractions) {
             String placeName = searchApiParser.extractNameFromLocation(attraction);
             XY xy = searchApiParser.extractXYFromLocation(attraction);
-            String imageUrl = imageApiManager.findPlaceImage(placeName, xy.x(), xy.y());
+            String imageUrl = locationService.findPlaceImage(placeName, xy);
             ((ObjectNode) attraction).put("image_url", imageUrl);
         }
         return new AttractionGetResponse(randomAttractions);
